+++ src/FileOp.cc
@@ -2,7 +2,6 @@
 #include "Base.hpp"
 #include "OptProcessor.hpp"
 #include <fmt/std.h>
-#include <cstddef>
 #include <fstream>
 #include <filesystem>
 #include <string>
@@ -15,10 +14,10 @@ namespace fs = std::filesystem;
 void readFromPath(const fs::path& path, std::string& str) {
   std::ifstream ifs{path, std::fstream::binary};
   if(!ifs) exitWithErr("Unable to open {} for reading", path);
-  size_t fsize{fs::file_size(path)};
+  uintmax_t fsize{fs::file_size(path)};
 #ifdef __cpp_lib_string_resize_and_overwrite
   str.resize_and_overwrite(fsize, [fsize]
-    ([[maybe_unused]] char* _, [[maybe_unused]] size_t _1) {
+    ([[maybe_unused]] char* _, [[maybe_unused]] uintmax_t _1) {
     return fsize;
   });
 #else
